pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "microfarm"
version = "0.1"
authors = [
    {name = "Souheil Chelfouh", email = "sch@treegital.fr"},
]
requires-python = ">=3.9"
classifiers = ["License :: OSI Approved :: MIT License"]
dependencies = [
  "aiozmq",
  "pydantic",
  "pydantic[email]",
  "pyjwt",
  "sanic",
  "sanic[ext]",
]

[tool.setuptools.packages.find]
where = ["."]
